+++ src/Entity/StationMedia.php
@@ -327,7 +327,7 @@ class StationMedia
         $length_min = floor($length / 60);
         $length_sec = $length % 60;
 
-        $this->length = round($length);
+        $this->length = (int)round($length);
         $this->length_text = $length_min . ':' . str_pad($length_sec, 2, '0', STR_PAD_LEFT);
     }
